+++ evm/contracts/proto/ibc/applications/fee/v1/ack.sol
@@ -1,4 +1,5 @@
 pragma solidity ^0.8.23;
+
 import "../../../../ProtoBufRuntime.sol";
 import "../../../../GoogleProtobufAny.sol";
 
@@ -18,7 +19,7 @@ library IbcApplicationsFeeV1IncentivizedAcknowledgement {
      * @return The decoded struct
      */
     function decode(bytes memory bs) internal pure returns (Data memory) {
-        (Data memory x, ) = _decode(32, bs, bs.length);
+        (Data memory x,) = _decode(32, bs, bs.length);
         return x;
     }
 
@@ -28,7 +29,7 @@ library IbcApplicationsFeeV1IncentivizedAcknowledgement {
      * @param bs The bytes array to be decoded
      */
     function decode(Data storage self, bytes memory bs) internal {
-        (Data memory x, ) = _decode(32, bs, bs.length);
+        (Data memory x,) = _decode(32, bs, bs.length);
         store(x, self);
     }
 
@@ -46,7 +47,7 @@ library IbcApplicationsFeeV1IncentivizedAcknowledgement {
         uint256 p,
         bytes memory bs,
         uint256 sz
-    ) internal pure returns (Data memory, uint) {
+    ) internal pure returns (Data memory, uint256) {
         Data memory r;
         uint256 fieldId;
         ProtoBufRuntime.WireType wireType;
@@ -54,10 +55,8 @@ library IbcApplicationsFeeV1IncentivizedAcknowledgement {
         uint256 offset = p;
         uint256 pointer = p;
         while (pointer < offset + sz) {
-            (fieldId, wireType, bytesRead) = ProtoBufRuntime._decode_key(
-                pointer,
-                bs
-            );
+            (fieldId, wireType, bytesRead) =
+                ProtoBufRuntime._decode_key(pointer, bs);
             pointer += bytesRead;
             if (fieldId == 1) {
                 pointer += _read_app_acknowledgement(pointer, bs, r);
@@ -66,11 +65,8 @@ library IbcApplicationsFeeV1IncentivizedAcknowledgement {
             } else if (fieldId == 3) {
                 pointer += _read_underlying_app_success(pointer, bs, r);
             } else {
-                pointer += ProtoBufRuntime._skip_field_decode(
-                    wireType,
-                    pointer,
-                    bs
-                );
+                pointer +=
+                    ProtoBufRuntime._skip_field_decode(wireType, pointer, bs);
             }
         }
         return (r, sz);
@@ -89,7 +85,7 @@ library IbcApplicationsFeeV1IncentivizedAcknowledgement {
         uint256 p,
         bytes memory bs,
         Data memory r
-    ) internal pure returns (uint) {
+    ) internal pure returns (uint256) {
         (bytes memory x, uint256 sz) = ProtoBufRuntime._decode_bytes(p, bs);
         r.app_acknowledgement = x;
         return sz;
@@ -106,7 +102,7 @@ library IbcApplicationsFeeV1IncentivizedAcknowledgement {
         uint256 p,
         bytes memory bs,
         Data memory r
-    ) internal pure returns (uint) {
+    ) internal pure returns (uint256) {
         (string memory x, uint256 sz) = ProtoBufRuntime._decode_string(p, bs);
         r.forward_relayer_address = x;
         return sz;
@@ -123,7 +119,7 @@ library IbcApplicationsFeeV1IncentivizedAcknowledgement {
         uint256 p,
         bytes memory bs,
         Data memory r
-    ) internal pure returns (uint) {
+    ) internal pure returns (uint256) {
         (bool x, uint256 sz) = ProtoBufRuntime._decode_bool(p, bs);
         r.underlying_app_success = x;
         return sz;
@@ -158,47 +154,32 @@ library IbcApplicationsFeeV1IncentivizedAcknowledgement {
         Data memory r,
         uint256 p,
         bytes memory bs
-    ) internal pure returns (uint) {
+    ) internal pure returns (uint256) {
         uint256 offset = p;
         uint256 pointer = p;
 
         if (r.app_acknowledgement.length != 0) {
             pointer += ProtoBufRuntime._encode_key(
-                1,
-                ProtoBufRuntime.WireType.LengthDelim,
-                pointer,
-                bs
+                1, ProtoBufRuntime.WireType.LengthDelim, pointer, bs
             );
             pointer += ProtoBufRuntime._encode_bytes(
-                r.app_acknowledgement,
-                pointer,
-                bs
+                r.app_acknowledgement, pointer, bs
             );
         }
         if (bytes(r.forward_relayer_address).length != 0) {
             pointer += ProtoBufRuntime._encode_key(
-                2,
-                ProtoBufRuntime.WireType.LengthDelim,
-                pointer,
-                bs
+                2, ProtoBufRuntime.WireType.LengthDelim, pointer, bs
             );
             pointer += ProtoBufRuntime._encode_string(
-                r.forward_relayer_address,
-                pointer,
-                bs
+                r.forward_relayer_address, pointer, bs
             );
         }
         if (r.underlying_app_success != false) {
             pointer += ProtoBufRuntime._encode_key(
-                3,
-                ProtoBufRuntime.WireType.Varint,
-                pointer,
-                bs
+                3, ProtoBufRuntime.WireType.Varint, pointer, bs
             );
             pointer += ProtoBufRuntime._encode_bool(
-                r.underlying_app_success,
-                pointer,
-                bs
+                r.underlying_app_success, pointer, bs
             );
         }
         return pointer - offset;
@@ -217,7 +198,7 @@ library IbcApplicationsFeeV1IncentivizedAcknowledgement {
         Data memory r,
         uint256 p,
         bytes memory bs
-    ) internal pure returns (uint) {
+    ) internal pure returns (uint256) {
         /**
          * First encoded `r` into a temporary array, and encode the actual size used.
          * Then copy the temporary array into `bs`.
@@ -242,14 +223,11 @@ library IbcApplicationsFeeV1IncentivizedAcknowledgement {
      * @param r The struct to be encoded
      * @return The number of bytes encoded in estimation
      */
-    function _estimate(Data memory r) internal pure returns (uint) {
+    function _estimate(Data memory r) internal pure returns (uint256) {
         uint256 e;
         e += 1 + ProtoBufRuntime._sz_lendelim(r.app_acknowledgement.length);
-        e +=
-            1 +
-            ProtoBufRuntime._sz_lendelim(
-                bytes(r.forward_relayer_address).length
-            );
+        e += 1
+            + ProtoBufRuntime._sz_lendelim(bytes(r.forward_relayer_address).length);
         e += 1 + 1;
         return e;
     }
